+++ NEWS.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `GenericAssemblyStrategy`. Since PR [#654](https://github.com/gridap/Gridap.jl/pull/654).
+
 ### Fixed
 - Inheritance relationship for DiscreteModelPortion. Since PR [#645](https://github.com/gridap/Gridap.jl/pull/645).
 - Optimization to RT FEs. Since PR [#638](https://github.com/gridap/Gridap.jl/pull/638).